third_party/WebKit/LayoutTests/crypto/aes-export-key-expected.txt

Test exporting an AES key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


error is: TypeError: Failed to execute 'exportKey' on 'SubtleCrypto': 2 arguments required, but only 1 present.
error is: DataError: Invalid key argument
error is: DataError: Invalid key argument
error is: DataError: Invalid key argument
error is: DataError: Invalid key argument

Importing a JWK key...
error is: SyntaxError: Invalid keyFormat argument
error is: SyntaxError: Invalid keyFormat argument
error is: SyntaxError: Invalid keyFormat argument
error is: SyntaxError: Invalid keyFormat argument
error is: SyntaxError: Invalid keyFormat argument

Exporting the key as raw data...
PASS bytesToHexString(new Uint8Array(exportedData)) is '8e73b0f7da0e6452c810f32b809079e562f8ead2522c6b7b'
Exporting the key as JWK...
PASS exportedJWK.kty is 'oct'
PASS exportedJWK.k is 'jnOw99oOZFLIEPMrgJB55WL46tJSLGt7'
PASS exportedJWK.alg is 'A192CBC'
PASS exportedJWK.ext is true
PASS exportedJWK.use is undefined
PASS exportedJWK.key_ops is ['encrypt', 'decrypt', 'wrapKey', 'unwrapKey']

Importing a key that's not extractable...

Trying to export as raw...
PASS Rejected, as expected
error is: InvalidAccessError: key is not extractable
Trying to export as jwk...
PASS Rejected, as expected
error is: InvalidAccessError: key is not extractable
PASS successfullyParsed is true

TEST COMPLETE
